enhanced LocationManager, better naming for read, for best practice enhanced code blocks, fixed UI, added imageloader and caching

+++ WeatherApp/Models/WeatherResponse.swift
@@ -38,7 +38,7 @@ struct Coord: Codable {
 // MARK: - Main
 struct Main: Codable {
     let temp: Double?
-    let feelsLike: Int?
+    let feelsLike: Double?
     let tempMin, tempMax: Double?
     let pressure, humidity, seaLevel, grndLevel: Int?
